# Conflicts:
#	lib/blocs/refresh_bloc_mixin.dart
#	lib/blocs/wallet_connect/chains/nom_service.dart
#	lib/l10n/app_en.arb
#	lib/main.dart
#	lib/main_dev.dart
#	lib/rearchitecture/features/features.dart
#	lib/rearchitecture/features/latest_transactions/bloc/latest_transactions_bloc.dart
#	lib/rearchitecture/features/latest_transactions/view/latest_transactions_card.dart
#	lib/rearchitecture/features/pending_transactions/bloc/pending_transactions_bloc.dart
#	lib/rearchitecture/features/pending_transactions/view/pending_transactions_card.dart
#	lib/rearchitecture/features/receive/view/receive_card.dart
#	lib/rearchitecture/features/receive/widgets/receive_populated.dart
#	lib/rearchitecture/features/send/bloc/send_transaction/send_transaction_bloc.dart
#	lib/rearchitecture/features/send/bloc/send_transaction/send_transaction_event.dart
#	lib/rearchitecture/features/send/view/send_card.dart
#	lib/rearchitecture/features/send/widgets/send_populated.dart
#	lib/rearchitecture/utils/blocs/infinite_list/infinite_list_bloc.dart
#	lib/rearchitecture/utils/constants/app_sizes.dart
#	lib/rearchitecture/utils/exceptions/exceptions.dart
#	lib/rearchitecture/utils/functions/functions.dart
#	lib/rearchitecture/utils/models/card/card_type.dart
#	lib/rearchitecture/utils/widgets/dropdowns/new_addresses_dropdown.dart
#	lib/rearchitecture/utils/widgets/dropdowns/zts_dropdown.dart
#	lib/rearchitecture/utils/widgets/infinite_scroll_table/cells/asset_cell.dart
#	lib/rearchitecture/utils/widgets/infinite_scroll_table/infinite_scroll_table_cell.dart
#	lib/utils/zts_utils.dart
#	lib/widgets/main_app_container.dart
#	lib/widgets/modular_widgets/pillar_widgets/pillar_collect.dart
#	lib/widgets/modular_widgets/settings_widgets/addresses.dart
#	lib/widgets/reusable_widgets/buttons/copy_to_clipboard_button.dart
#	lib/widgets/reusable_widgets/receive_qr_image.dart
#	pubspec.lock
#	pubspec.yaml
#	test/multiple_balance/multiple_balance_bloc_test.dart
#	test/send/bloc/send_transaction_bloc_test.dart

+++ lib/widgets/modular_widgets/p2p_swap_widgets/p2p_swaps_card.dart
@@ -8,11 +8,11 @@ import 'package:zenon_syrius_wallet_flutter/widgets/modular_widgets/p2p_swap_wid
 import 'package:zenon_syrius_wallet_flutter/widgets/widgets.dart';
 
 class P2pSwapsCard extends StatefulWidget {
-
   const P2pSwapsCard({
     required this.onStepperNotificationSeeMorePressed,
     super.key,
   });
+
   final VoidCallback onStepperNotificationSeeMorePressed;
 
   @override
@@ -99,31 +99,35 @@ class _P2pSwapsCardState extends State<P2pSwapsCard> {
   }
 
   Future<void> _onDeleteSwapTapped(P2pSwap swap) async {
-    showDialogWithNoAndYesOptions(
-        context: context,
-        isBarrierDismissible: true,
-        title: 'Delete swap',
-        description:
-            'Are you sure you want to delete this swap? This action cannot be undone.',
-        onYesButtonPressed: () async {
-          if (swap.mode == P2pSwapMode.htlc) {
-            await htlcSwapsService!.deleteSwap(swap.id);
-          }
-          _p2pSwapsListBloc.getData();
-        },);
+    final bool? deleteConfirmed = await showDialogWithNoAndYesOptions(
+      context: context,
+      isBarrierDismissible: true,
+      title: 'Delete swap',
+      description:
+          'Are you sure you want to delete this swap? This action cannot be undone.',
+    );
+
+    if (deleteConfirmed ?? false) {
+      if (swap.mode == P2pSwapMode.htlc) {
+        await htlcSwapsService!.deleteSwap(swap.id);
+      }
+      _p2pSwapsListBloc.getData();
+    }
   }
 
   Future<void> _onDeleteSwapHistoryTapped() async {
-    showDialogWithNoAndYesOptions(
-        context: context,
-        isBarrierDismissible: true,
-        title: 'Delete swap history',
-        description:
-            'Are you sure you want to delete your swap history? Active swaps cannot be deleted.',
-        onYesButtonPressed: () async {
-          await htlcSwapsService!.deleteInactiveSwaps();
-          _p2pSwapsListBloc.getData();
-        },);
+    final bool? deleteHistoryConfirmed =  await showDialogWithNoAndYesOptions(
+      context: context,
+      isBarrierDismissible: true,
+      title: 'Delete swap history',
+      description:
+          'Are you sure you want to delete your swap history? Active swaps cannot be deleted.',
+    );
+
+    if (deleteHistoryConfirmed ?? false) {
+      await htlcSwapsService!.deleteInactiveSwaps();
+      _p2pSwapsListBloc.getData();
+    }
   }
 
   Widget _getTable(List<P2pSwap> swaps) {
@@ -143,22 +147,23 @@ class _P2pSwapsCardState extends State<P2pSwapsCard> {
             child: Scrollbar(
               controller: _scrollController,
               child: ListView.separated(
-                  controller: _scrollController,
-                  cacheExtent: 1000,
-                  itemCount: swaps.length,
-                  separatorBuilder: (_, __) {
-                    return const SizedBox(
-                      height: 15,
-                    );
-                  },
-                  itemBuilder: (_, int index) {
-                    return P2pSwapsListItem(
-                      key: ValueKey(swaps.elementAt(index).id),
-                      swap: swaps.elementAt(index),
-                      onTap: _onSwapTapped,
-                      onDelete: _onDeleteSwapTapped,
-                    );
-                  },),
+                controller: _scrollController,
+                cacheExtent: 1000,
+                itemCount: swaps.length,
+                separatorBuilder: (_, __) {
+                  return const SizedBox(
+                    height: 15,
+                  );
+                },
+                itemBuilder: (_, int index) {
+                  return P2pSwapsListItem(
+                    key: ValueKey(swaps.elementAt(index).id),
+                    swap: swaps.elementAt(index),
+                    onTap: _onSwapTapped,
+                    onDelete: _onDeleteSwapTapped,
+                  );
+                },
+              ),
             ),
           ),
         ],
